move context, store_artifact, store_id to icp crate
