+++ crates/icp-cli/src/commands/canister/settings/update.rs
@@ -6,11 +6,12 @@ use ic_agent::{AgentError, export::Principal};
 use ic_management_canister_types::{CanisterStatusResult, EnvironmentVariable, LogVisibility};
 use icp::{agent, identity, network};
 
-use crate::{
-    commands::Context,
-    options::{EnvironmentOpt, IdentityOpt},
+use icp::{
+    context::Context,
     store_id::{Key, LookupError as LookupIdError},
 };
+
+use crate::options::{EnvironmentOpt, IdentityOpt};
 
 #[derive(Clone, Debug, Default, Args)]
 pub(crate) struct ControllerOpt {
