+++ crates/icp-cli/src/commands/canister/stop.rs
@@ -2,11 +2,12 @@ use clap::Args;
 use ic_agent::AgentError;
 use icp::{agent, identity, network};
 
-use crate::{
-    commands::Context,
-    options::{EnvironmentOpt, IdentityOpt},
+use icp::{
+    context::Context,
     store_id::{Key, LookupError as LookupIdError},
 };
+
+use crate::options::{EnvironmentOpt, IdentityOpt};
 
 #[derive(Debug, Args)]
 pub(crate) struct StopArgs {
