+++ crates/icp-cli/src/commands/environment/list.rs
@@ -1,6 +1,6 @@
 use clap::Args;
 
-use crate::commands::Context;
+use icp::context::Context;
 
 #[derive(Debug, Args)]
 pub(crate) struct ListArgs;
