+++ crates/icp-cli/src/commands/identity/principal.rs
@@ -1,7 +1,7 @@
 use clap::Args;
-use icp::identity;
+use icp::{context::Context, identity};
 
-use crate::{commands::Context, options::IdentityOpt};
+use crate::options::IdentityOpt;
 
 #[derive(Debug, Args)]
 pub(crate) struct PrincipalArgs {
