+++ crates/icp-cli/src/commands/network/stop.rs
@@ -1,10 +1,8 @@
 use std::time::Duration;
 
 use clap::Parser;
-use icp::{fs::remove_file, manifest, network::NetworkDirectory};
+use icp::{context::Context, fs::remove_file, manifest, network::NetworkDirectory};
 use sysinfo::{Pid, ProcessesToUpdate, Signal, System};
-
-use crate::commands::Context;
 
 const TIMEOUT_SECS: u64 = 30;
